VBM-159: required changes applied
