+++ packages/mobile/src/screens/items-and-services/components/components.ts
@@ -1,1 +1,2 @@
 export { Filter } from './filter/filter';
+export { ListHeader } from './list-header/list-header';
